Break down file
