+++ packages/alfa-media/src/condition/index.ts
@@ -1,1 +1,4 @@
+export * from "./and";
 export * from "./condition";
+export * from "./not";
+export * from "./or";
