+++ packages/alfa-media/tsconfig.json
@@ -2,8 +2,11 @@
   "$schema": "http://json.schemastore.org/tsconfig",
   "extends": "../tsconfig.json",
   "files": [
+    "src/condition/and.ts",
     "src/condition/condition.ts",
     "src/condition/index.ts",
+    "src/condition/not.ts",
+    "src/condition/or.ts",
     "src/feature/feature.ts",
     "src/feature/index.ts",
     "src/index.ts",
